refactor(core): extract ChartViewportManager from Chart
Move DPR management, ResizeObserver, scroll synchronization, viewport
computation, and viewport signal into dedicated ChartViewportManager.

- chart.ts: -208 lines (2131 -> 1923)
- chartViewportManager.ts: 310 lines (new)
- All public API signatures preserved via thin delegation
- All 48 test files (956 tests) pass
